[libc++] Refactor locale_guard
Rename __libcpp_locale_guard to just __locale_guard, since there's no
reason for it to have __libcpp_ in its name -- it's just an internal
utility.

Also, define __locale_guard unconditionally of _LIBCPP_LOCALE__L_EXTENSIONS,
since that header is only used on Windows (where it has a custom definition)
or from bsd_locale_fallbacks.h, which is only included when the L extensions
are not provided.
